extract Wikidata source class
Split Source::Wikidata out into its own file
