+++ lib/source.rb
@@ -98,19 +98,6 @@ module Source
     end
   end
 
-
-  class Wikidata < Person
-    def fields
-      super << :identifier__wikidata
-    end
-
-    def reconciliation_file
-      Reconciliation::File.new(
-        Pathname.new('sources/') + i(:merge)[:reconciliation_file]
-      )
-    end
-  end
-
   class OCD < CSV
     def fields
       %i(area area_id)
